change drone entity id type

+++ src/main/java/com/interview/drone/backend/repository/DroneRepository.java
@@ -4,6 +4,8 @@ import com.interview.drone.backend.entity.Drone;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
-public interface DroneRepository extends JpaRepository<Drone, Long> {
+public interface DroneRepository extends JpaRepository<Drone, BigInteger> {
 }
